requirements.txt
requests
beautifulsoup4
pyyaml
openai
playwright
